Several Flax Improvements
Now does a better job at running through flax in the order that it's
planted, operates a little faster. Works on a 4x4 planting grid and 6x6
seed grid (reliably with 4 seeds per harvest). Additionally, added an
input box to choose which type of Flax to use at start time.