Several macro updates
Updated cc_Assist, cc_Auto, gather and wood macro's to work with T8. Had to add new cc images as it refused to work with old ones even when updated. Added dirt image for statclicks as it was missing. Updated images needed for wood macro. Created new default routes for gather macro for use in T8.