Update glass.lua
Macro was clicking outside the window when clicking to refresh. Changed a -10 x coord to a +10 to make sure it clicks inside the bounds. Added a y offset to ensure it doesn't click a submenu. Specific number for now, may change it to just be like 90% of window height later.