Add: veggies_semiAuto.lua
Assists in growing veggies. Unlike veggies.lua which simply only plants
seeds and quits; this macro is put in a loop. You plant and pin the
veggies (optionally includes windowManager to assist in pinning and
placing in grids).

Once all your windows are pinned, it will pause and wait for you to
press the Trigger Key (Shift, Alt, Ctrl or MouseWheel click). Each time,
it will water your veggies. If the pinned windows shows Harvest, it will
harvest them.