test_ocr_clock.lua -- More Robust
Script can now detect if you are running the new T8 Clock or Legacy. It
can also detect which Interface Size (of either Legacy or T8 Clock)

This is should be a great diagnostic tool to report to player if they
are running New T8 Clock and/or if they are running the wrong Interface
Size.