mining_ore.lua -- See github comments
v2.1.0 -- Attempt to deal with mining in regions where this a faction
bonus of +##% bonus to yield.

This causes double messages to appear in chat and potentially screw up
the macro.   New routine will ignore it it happens to capture the bonus
message and wait for next valid message.  Some Misc. tweaks, also ...