distillationCoil.lua
Removed 2000 tolerance so it will use default of 5000.

Includes common.inc only now.

Added statusScreen to report the keystrokes it's sending to the glory hole. Previously, after it said "clicking" it would just hang (while sending keystrokes) to confuse player that it might not be doing anything (when in fact it was).