mining_ore.lua v2.2.0
Checkbox added "Also do 4 stone combinations"

Rarely you will get ore when using 4 stone combinations, but if you do, it can be 100-200+ per hit.

The option is there for you to decide if it's worth it or not.

If you choose to do 4 stone combination, it will do all the 4 stone combinations FIRST, then do all the 3 stone combos (as usual).

So beware, this will double the time it takes to work the mine each round. It's a gamble. But the option is there for you to decide.