fishing.lua v2.0.7 -- See Desc
Prevent issues of not finding more than 30 lures, when player has more than one window pinned.

If player had 2 windows (Lure Menu and Fillet Menu) and if the Fillet Menu was on the left side of Lure window; then it'd find the Fillet Menu first.

When this happens, the srFindImageInRange to find down arrow (on scroll menu) fails to be found and skipped .. Because it's looking on the wrong pinned menu.